Restore original functionality in safari and in brave
